Refactor PetControllerTest
