+++ src/test/java/pro/sky/animal_shelter_telegram_bot/controller/PetOwnerControllerTests/PetOwnerControllerTest.java
@@ -4,7 +4,6 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.client.TestRestTemplate;
-import org.springframework.boot.test.web.server.LocalServerPort;
 import pro.sky.animal_shelter_telegram_bot.controller.PetOwnerController;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -13,8 +12,7 @@ import static pro.sky.animal_shelter_telegram_bot.controller.ConstantsForControl
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 class PetOwnerControllerTest {
 
-    @LocalServerPort
-    private int port;
+    private final String LOCAL_URL = URL + PORT + "/" + PET_OWNER_URL + "/";
 
     @Autowired
     private PetOwnerController petOwnerController;
@@ -29,11 +27,11 @@ class PetOwnerControllerTest {
 
     @Test
     public void testHelloMessage(){
-        assertThat(this.restTemplate.getForObject(URL + port + "/" + PET_OWNER_URL + "/", String.class))
+        assertThat(this.restTemplate.getForObject(LOCAL_URL, String.class))
                 .isNotEmpty();
-        assertThat(this.restTemplate.getForObject(URL + port + "/" + PET_OWNER_URL + "/", String.class))
+        assertThat(this.restTemplate.getForObject(LOCAL_URL, String.class))
                 .isNotNull();
-        assertThat(this.restTemplate.getForObject(URL + port + "/" + PET_OWNER_URL + "/", String.class))
+        assertThat(this.restTemplate.getForObject(LOCAL_URL, String.class))
                 .isEqualTo(HELLO_MESSAGE_PET_OWNER);
     }
 }
